+++ src/shared/components/SubmissionManagement/SubmissionManagement/index.jsx
@@ -26,7 +26,6 @@ import style from './styles.scss';
 
 export default function SubmissionManagement(props) {
   const {
-    auth,
     challenge,
     submissions,
     loadingSubmissions,
@@ -157,7 +156,6 @@ export default function SubmissionManagement(props) {
         {!loadingSubmissions
           && (
           <SubmissionsTable
-            auth={auth}
             submissionObjects={submissions}
             showDetails={showDetails}
             track={track}
@@ -210,5 +208,4 @@ SubmissionManagement.propTypes = {
   loadingSubmissions: PT.bool,
   challengeUrl: PT.string,
   submissionPhaseStartDate: PT.string.isRequired,
-  auth: PT.shape().isRequired,
 };
